🎨 文言などを提示する関数を別ファイルに切り分け
保守性を上げるため
